showing/hiding albumRoll
those menus will show when menu is not moving, and hide when switching from one to another
changeSelection now also has a boolean argument of instant showing or 0.25 delay. different action will use as false or off

Also makes updateAlbum() in AlbumRoll public instead of none, that is just easier!